requirements.txt
flask
flask-bcrypt
flask-socketio
flask-sqlalchemy
flask-limiter
flask-migrate
pymysql
python-dotenv
watchdog
gunicorn
eventlet